chore: watch theme state in app and set attribute on root
